% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.r
\name{credentials}
\alias{credentials}
\title{Author: Sarkis Jacob
CRC/OTTAWA - Canada}
\usage{
credentials(username, passwd)
}
\arguments{
\item{username, passwd}{}
}
\value{
username and password
}
\description{
January 29, 2019 --
This program registers user's credentials
}
\examples{
credentials(user,password)
}
