% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisconnectSpark.r
\name{disconnect_spark}
\alias{disconnect_spark}
\title{Author: Sarkis Jacob
CRC/OTTAWA - Canada}
\usage{
disconnect_spark()
}
\value{
Disconnects spark
}
\description{
January 29, 2019 --
This program will disconnect spark
}
\examples{

}
