% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GpsCoordinatesWithinBoundaryBox.r
\name{gps_coordinates_within_box}
\alias{gps_coordinates_within_box}
\title{Author: Sarkis Jacob
CRC/OTTAWA - Canada}
\usage{
gps_coordinates_within_box(latitude, longitude, box_length_km)
}
\arguments{
\item{latitude}{Defines the latitude coordinate for the center of the boundary-box containing the dataset.}

\item{longitude}{Defines the longitude coordinate for the center of the boundary-box containing the dataset.}

\item{box_length_km}{Defines the side-length of the boundary-box containing the dataset.}
}
\value{
Spark dataframe
}
\description{
January 29, 2019 --
This program filters a dataset within a bounary-box centered at coordinates (latitude and longitude) and with side-length (box_length_km) specified by the user.
}
\examples{
gps_coordinates_within_box(latitude,longitude, box_length_km)
}
