% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GpsCoordinatesWithinBoundaryCircle.r
\name{gps_coordinates_within_circle}
\alias{gps_coordinates_within_circle}
\title{Author: Sarkis Jacob
CRC/OTTAWA - Canada}
\usage{
gps_coordinates_within_circle(latitude, longitude, circle_radius_km)
}
\arguments{
\item{latitude}{Defines the latitude coordinate for the center of the circle containing the dataset.}

\item{longitude}{Defines the longitude coordinate for the center the circle containing the dataset.}

\item{circle_radius_km}{Defines the radius coordinate of the circle containing the dataset.}
}
\value{
Spark dataframe
}
\description{
January 29, 2019 --
This program filters a dataset within a bounary-circle centered at coordinates (latitude and longitude) and with radius (circle_radius_km) specified by the user.
}
\examples{
gps_coordinates_within_circle(latitude,longitude, circle_radius_km)
}
